moved the merging of analysis and tweets from the tweets_data_analysis to the tone_analyzer class because it seems like a better fit. Added a method that will grab the emotion per month, right now it just displays the tweet with the most favorites.
